+++ packages/cli/src/commands/commonOptions.ts
@@ -1,5 +1,4 @@
 export interface CommonOptions {
   config?: string
   skipCodegen?: boolean
-  descriptorsPackage?: boolean
 }