+++ packages/client/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- cli: Option flag `--no-descriptors-package` to generate descriptors without installing them as a package.
+- cli: Option flag `noDescriptorsPackage` to generate descriptors without installing them as a package.
 
 ## 1.7.3 - 2024-11-08